backend/core/requirements.txt
APScheduler==3.10.4
asyncpg==0.29.0
aiohttp==3.10.5
aiolimiter==1.1.0
alembic==1.14.0
beautifulsoup4==4.12.2
fastapi==0.111.0
icecream==2.1.3
psycopg==3.1.19
psycopg-binary==3.2.6
psycopg2==2.9.10
pydantic==2.7.3
pydantic-settings==2.3.1
lxml==5.3.0
requests==2.31.0
requests_oauthlib==2.0.0
sqlalchemy==2.0.30
# selenium==4.11.2
serpapi==0.1.5
unicorn==2.0.1.post1

authx~=1.4.2
watchfiles~=1.0.5
python-dotenv~=1.1.0
uvicorn~=0.34.0
starlette~=0.37.2
pytz~=2025.2

coloredlogs==15.0.1
colorlog==6.9.0
pyjwt==2.10.1
bcrypt==4.3.0